src/server/mcp_hub/whatsapp/requirements.txt
fastmcp
python-dotenv
motor
cryptography
httpx
